+++ CHANGELOG.md
@@ -8,6 +8,7 @@
   - AppleALC 1a3e5cb (1.6.0 rolling - 04-10-2021)
 - Enhance Wifi model detection
 - Hide OpenShell.efi by default
+- Add Brightness Control patches for legacy Nvidia and Intel GPUs
 
 ## 0.0.22
 - Add ExFat support for models missing driver